Consolidate test_forward and test_forward2
